MachineLearningBasicAPP
This app follows a tutorial off of PluralSight. This tutorial was very informative and good for programmers wanting to have an introductory understanding of machine learning and python. This app reads data from an excel file, then is able to find correlations within that data. Also, this app does include some visualization for the lines in the python that are "img". There is a JPEG that shows that visualization. The GaussianNB algorithm helps with predictions throughout this application.